+++ src/main/java/cn/elytra/mod/nomi_horizons/utils/reflection/FieldAccessor.java
@@ -1,0 +1,7 @@
+package cn.elytra.mod.nomi_horizons.utils.reflection;
+
+public interface FieldAccessor<T> {
+    T get(Object self);
+
+    void set(Object self, T value);
+}